feat(place) : ServiceImpl 패키지 재생성
