+++ src/main/java/kr/co/mapspring/place/service/AdminMissionService.java
@@ -1,38 +1,8 @@
 package kr.co.mapspring.place.service;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import kr.co.mapspring.global.exception.place.ScenarioNotFoundException;
 import kr.co.mapspring.place.dto.AdminCreateMissionDto;
-import kr.co.mapspring.place.entity.Mission;
-import kr.co.mapspring.place.entity.Scenario;
-import kr.co.mapspring.place.repository.MissionRepository;
-import kr.co.mapspring.place.repository.ScenarioRepository;
-import lombok.RequiredArgsConstructor;
 
-@Service
-@RequiredArgsConstructor
-public class AdminMissionService {
+public interface AdminMissionService {
 	
-	private final ScenarioRepository scenarioRepository;
-	private final MissionRepository missionRepository;
-	
-	// 미션 생성
-	@Transactional
-	public void createMission (AdminCreateMissionDto.RequestCreate request) {
-		
-		Long scenarioId = request.getScenarioId();
-		
-		Scenario scenario = scenarioRepository.findById(scenarioId)
-				.orElseThrow(ScenarioNotFoundException::new);
-		
-		Mission mission = Mission.create(request.getMissionTitle(),
-										 request.getMissionDescription(),
-										 scenario);
-		
-		missionRepository.save(mission);
-		
-	}
-    
+	void createMission (AdminCreateMissionDto.RequestCreate request);
 }
